+++ CHANGELOG.md
@@ -20,6 +20,13 @@
 
 - bumped dependency `rand = "0.3.18"`.
 
+- Added `proptest::collection::subsequence<T: Clone>(Vec<T>, Range<usize>)`
+  which returns a strategy generating subsequences, of the source `Vec`,
+  with a size within the given `Range`.
+
+- Added the trait `CollectionStrategy` which provides `.prop_sample` directly
+  on collection types. The operation is the same as `subsequence` for `Vec`.
+
 ## 0.3.1
 
 ### New Additions